update the values.yaml file with the following changes
To improve the Helm chart, we can make the following changes:

Add a check for values.yaml during chart install:

If no storage is specified, the chart should use an EmptyDir volume for the application to start.
If storage is specified, the chart should create a PersistentVolumeClaim (PVC) for the application.
Allow specifying the storageClass via chart reconfiguration:

We can add a configurable parameter to the chart to allow users to change the storageClass for the PVC created by the chart.
Respect the specified storageClass:

If a storageClass is specified in the values.yaml, the chart should create a PVC with the specified storageClass.
Check for the presence of hostPath and storageClass:

If a hostPath is specified without a storageClass, the chart should not create a PVC and use the specified hostPath instead.
Add appropriate error handling:

If there is an error with creating the PVC, the chart should fail with an appropriate error message.
By implementing these changes, we can ensure that the Helm chart will work in different scenarios, including those where no storage is specified or where users want to change the storageClass after installation. Additionally, we can provide appropriate error handling to make it easier for users to diagnose and fix any issues that may arise during installation.